+++ changelog.md
@@ -9,6 +9,7 @@ _Date TBD_
 
 **Performance Improvements:**
 
+* Update `plus` to concatenate Immutable Arrays without using builders.
 * Update `sortedWith` on `ImmutableArray<T>` to copy the elements with `arraycopy`.
 * Update `partition` to use `arraycopy` to copy from the left side of the buffer.